% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_fert_rates.R
\name{act_fert_rates}
\alias{act_fert_rates}
\title{Actual N inputs from fertilizers}
\usage{
act_fert_rates(data, databases)
}
\arguments{
\item{data}{Field data}

\item{databases}{Databases}
}
\value{
data_act N inputs from fertilizers
}
\description{
Actual N inputs from fertilizers
}
\examples{
A
}
