% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_n_fixing.R
\name{act_n_fixing}
\alias{act_n_fixing}
\title{N-fixing species}
\usage{
act_n_fixing(data, databases)
}
\arguments{
\item{data}{Field dataset}

\item{databases}{Databases include necessary default values for the calculation}
}
\value{
Results of N2O emissions due to the use of N-fixing species
}
\description{
N-fixing species
}
