% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_n_inputs.R
\name{act_n_inputs}
\alias{act_n_inputs}
\title{N inputs from synthetic and organic fertilizers in the actual harvest year}
\usage{
act_n_inputs(data, databases)
}
\arguments{
\item{data}{Field dataset}

\item{databases}{Databases include necessary default values for the calculation}
}
\value{
N inputs from synthetic and organic fertilizers in the actual harvest year
}
\description{
N inputs from synthetic and organic fertilizers in the actual harvest year
}
