% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsl_fert_rates.R
\name{bsl_fert_rates}
\alias{bsl_fert_rates}
\title{N inputs in the baseline}
\usage{
bsl_fert_rates(data, databases)
}
\arguments{
\item{data}{field data}

\item{databases}{Databases}
}
\value{
data final data
}
\description{
N inputs in the baseline
}
\examples{
A
}
