% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{databases_2022}
\alias{databases_2022}
\title{Dataset 2: databases_2022}
\format{
A character
}
\source{
This dataset is included in the datasets package in R.
}
\usage{
data(databases_2022)
}
\description{
A path.
}
\details{
This dataset contains a path for databases for 2022 calculation.
}
\keyword{datasets}
