% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Dataset 1: example_data}
\format{
A data frame with 6 rows and 86 variables:
\describe{
\item{X}{id}
\item{field_qa_status}{status of field data quality control}
...
}
}
\source{
This dataset is included in the datasets package in R.
}
\usage{
data(example_data)
}
\description{
A data frame with 6 observations on 86 variables.
}
\details{
This dataset contains six examples of field data.
}
\keyword{datasets}
