% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2o_emissions_fert.R
\name{n2o_emissions_fert}
\alias{n2o_emissions_fert}
\title{Calculation for N2O emissions from soil}
\usage{
n2o_emissions_fert(data, scenario, databases)
}
\arguments{
\item{data}{field dataset}

\item{scenario}{Baseline or Actual harvest year}

\item{databases}{Databases include necessary default values for the calculation}
}
\value{
Results of N2O emissions from soil
}
\description{
Calculation for N2O emissions from soil
}
