% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2o_emissions_fert_cal.R
\name{n2o_emissions_fert_cal}
\alias{n2o_emissions_fert_cal}
\title{N2O emissions from soil}
\usage{
n2o_emissions_fert_cal(data, scenario, databases)
}
\arguments{
\item{data}{field data}

\item{scenario}{Baseline or Actual harvest year}

\item{databases}{Database}
}
\value{
data final data
}
\description{
N2O emissions from soil
}
\examples{
A
}
