% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_n_fixing.R
\name{act_n_fixing}
\alias{act_n_fixing}
\title{N-fixing species}
\usage{
act_n_fixing(data, databases)
}
\arguments{
\item{data}{field data}

\item{databases}{Databases}
}
\value{
Final data
}
\description{
N-fixing species
}
