% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_n_inputs.R
\name{act_n_inputs}
\alias{act_n_inputs}
\title{Actual N inputs from fertilizers}
\usage{
act_n_inputs(data, databases)
}
\arguments{
\item{data}{Field data}

\item{databases}{Databases}
}
\value{
data_act N inputs from fertilizers
}
\description{
Actual N inputs from fertilizers
}
\examples{
A
}
