% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgreenaIPCC.R
\name{AgreenaIPCC}
\alias{AgreenaIPCC}
\title{AgreenaIPPC model}
\usage{
AgreenaIPCC(data, harvest_year, databases)
}
\arguments{
\item{data}{Field data from the platform}

\item{harvest_year}{Define the target harvest year}

\item{databases}{Databases include necessary default values for the calculation}
}
\value{
data_bsl_act Final result table
}
\description{
AgreenaIPPC model
}
\examples{
library(readxl)
library(dplyr)
library(tidyr)
setwd("G:/My Drive/Work/Projects/Models/IPCC model/20240222_R models/3_Model")
data <- read.csv(file = "All_Baseline_field_data_aws_23.csv")
harvest_year <- 2022
databases <- "Databases.xlsx"
result <- AgreenaIPCC(data, harvest_year, databases)
}
