% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgreenaIPCC.R
\name{AgreenaIPCC}
\alias{AgreenaIPCC}
\title{AgreenaIPPC model}
\usage{
AgreenaIPCC(data, harvest_year, databases)
}
\arguments{
\item{data}{Field data from the platform}

\item{harvest_year}{Define the target harvest year}

\item{databases}{Databases include necessary default values for the calculation}
}
\value{
A final result table
}
\description{
AgreenaIPPC model
}
\examples{
library(readxl)
library(dplyr)
library(tidyr)
setwd("G:/My Drive/Work/Projects/Models/IPCC model/20240313_R package_2022")
data <- read.csv(file = "All_Baseline_field_data_aws_23.csv")
harvest_year <- 2022
databases <- "2022_Databases.xlsx"
result <- AgreenaIPCC(data, harvest_year, databases)

}
