% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgreenaIPCC.R
\name{AgreenaIPCC}
\alias{AgreenaIPCC}
\title{AgreenaIPPC model}
\usage{
AgreenaIPCC(data, harvest_year, databases)
}
\arguments{
\item{data}{Field dataset that farmer reported at the platform}

\item{harvest_year}{Define the target harvest year}

\item{databases}{Databases include necessary default values for the calculation}
}
\value{
A final IPCC result table including key results of baseline_fuel_emissions, actual_fuel_emissions, baseline_soil_n2o_emissions, actual_soil_n2o_emissions, baseline_lime_co2_emissions_(5y) and actual_lime_emissions
}
\description{
AgreenaIPPC model
}
\examples{
library(readxl)
library(dplyr)
library(tidyr)
setwd("G:/My Drive/Work/Projects/Models/IPCC model/20240314_R package_2023")
data <- read.csv(file = "All_Baseline_field_data_aws_23.csv")
harvest_year <- 2023
databases <- "2023_Databases.xlsx"
result <- AgreenaIPCC(data, harvest_year, databases)

}
