% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsl_n_inputs.R
\name{bsl_n_inputs}
\alias{bsl_n_inputs}
\title{N inputs in the baseline}
\usage{
bsl_n_inputs(data, databases)
}
\arguments{
\item{data}{field data}

\item{databases}{Databases}
}
\value{
data final data
}
\description{
N inputs in the baseline
}
\examples{
None
}
